+++ app/src/main/java/com/symeonchen/wakeupscreen/compose/AppInfoScreen.kt
@@ -56,7 +56,7 @@ fun AppInfoScreen(
 
             Surface(
                 shape = RoundedCornerShape(16.dp),
-                color = MaterialTheme.colorScheme.surface,
+                color = MaterialTheme.colorScheme.surfaceContainerLow,
                 tonalElevation = 0.dp,
                 modifier = Modifier.fillMaxWidth(),
             ) {