+++ app/src/main/java/com/symeonchen/wakeupscreen/compose/CheckUpdateScreen.kt
@@ -44,7 +44,7 @@ fun CheckUpdateScreen(
 
             Surface(
                 shape = RoundedCornerShape(16.dp),
-                color = MaterialTheme.colorScheme.surface,
+                color = MaterialTheme.colorScheme.surfaceContainerLow,
                 tonalElevation = 0.dp,
                 modifier = Modifier.fillMaxWidth(),
             ) {