+++ app/src/main/java/com/symeonchen/wakeupscreen/compose/FeedbackScreen.kt
@@ -31,7 +31,7 @@ fun FeedbackScreen(
         Column(modifier = Modifier.padding(16.dp)) {
             Surface(
                 shape = RoundedCornerShape(16.dp),
-                color = MaterialTheme.colorScheme.surface,
+                color = MaterialTheme.colorScheme.surfaceContainerLow,
                 tonalElevation = 0.dp,
                 modifier = Modifier.fillMaxWidth(),
             ) {